Data-Engineering/Day1/docker/requirements.txt
flask==2.0.1
werkzeug==2.0.1
SQLAlchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0 